requirements.txt
django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
django-environ==0.10.0